aggiunti many to many,seeder, validazione dati form

+++ app/Http/Controllers/Admin/PostController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin;
 
 use App\Post;
 use App\Category;
+use App\Tag;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Str;
@@ -31,7 +32,12 @@ class PostController extends Controller
     {
         // vado a prendermi tutte le categorie nel database e le passo alla view
         $categories = Category::all();
-        return view('admin.posts.create', ['categories'=> $categories]);
+        $tags = Tag::all();
+        return view('admin.posts.create', [
+            'categories'=> $categories,
+            'tags'=> $tags,
+        ]);
+
     }
 
     /**
@@ -42,27 +48,34 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
-        // prendo quello che mi arrivata dal form
+        $request->validate([
+            'title' =>'required|max:255',
+            'author' =>'required|max:255',
+            'content' =>'required',
+            'cover_image_file' =>'image'
+        ]);
+        // prendo quello che mi arrivata dal form(post)
         $dati= $request->all();
         // dd($dati);
-        // creo un entità e uso quello che è arrivato dal form per completarla con la funzione fill()
+        // creo un entità(oggetto) e uso quello che è arrivato dal form per completarla con la funzione fill()
         $post= new Post();
         $post->fill($dati);
 
         if (!empty($dati['cover_image_file'])) {
-            // prendo la parte che mi serve dall array
+            // prendo la parte che mi serve dall array, l utente ha caricato un immagine
             $cover_image = $dati['cover_image_file'];
             // uso la facades Storage con la funzione put,dicendogli dove mettere il file e cosa andare a prendere(il nostro oggettone),salvo tutto dentro una variabile visto che mi viene restituito il path di cui ho bisogno,questa put parte da storage\app\public,noi gli diciamo di creare la cartella uploads
             $cover_image_path = Storage::put('uploads', $cover_image);
             // assegno a mano questa parte, il resto viene compilato da fill()
             $post->cover_image= $cover_image_path;
         }
-
+        // recupero il titolo e genero lo slug corrispondente
         $slug_originale = Str::slug($dati['title']);
         $slug= $slug_originale;
         // verifico che nel db non esiste uno slug uguale
         $post_stesso_slug = Post::where('slug' , $slug)->first();
         $slug_trovati = 1;
+        // ciclo finche non trovo uno slug libero, se è falso subito non entro qui
         while (!empty($post_stesso_slug)) {
             $slug= $slug_originale . '-' . $slug_trovati;
             $post_stesso_slug = Post::where('slug' , $slug)->first();
@@ -70,8 +83,16 @@ class PostController extends Controller
         }
         $post->slug=$slug;
 
-
+        // salvo il post a db
         $post->save();
+        // creo i collegamenti nella tabella ponte, solo quelli selezionati, cancello gli altri, questo con la funzione sync che è molto utile per la funzione update
+
+
+        if (!empty($dati['tag_id'])) {
+            // sono stati selezionati dei tag, li assegno al post
+            $post->tags()->sync($dati['tag_id']);
+        }
+        // redirect homepage admin post
         return redirect()->route('admin.posts.index');
 
     }
@@ -96,9 +117,15 @@ class PostController extends Controller
      */
     public function edit(Post $post /*$id*/)
     {
+
         // $post = Post::find($id);
         $categories = Category::all();
-        return view('admin.posts.edit', ['post'=> $post , 'categories'=> $categories]);
+        $tags = Tag::all();
+        return view('admin.posts.edit', [
+            'post'=> $post ,
+            'categories'=> $categories,
+            'tags' => $tags
+        ]);
     }
 
     /**
@@ -110,10 +137,22 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        $request->validate([
+            'title' =>'required|max:255',
+            'author' =>'required|max:255',
+            'content' =>'required',
+            'cover_image_file' =>'image'
+        ]);
+        // recupero il post dal db
+        // $post = Post::find($id);
         // prendo quello che mi arrivata dal form
         $dati= $request->all();
 
         if (!empty($dati['cover_image_file'])) {
+            if (!empty($post->cover_image)) {
+                // cancello l immagine precedente
+                Storage::delete($post->cover_image);
+            }
             // prendo la parte che mi serve dall array
             $cover_image = $dati['cover_image_file'];
             // uso la facades Storage con la funzione put,dicendogli dove mettere il file e cosa andare a prendere(il nostro oggettone),salvo tutto dentro una variabile visto che mi viene restituito il path di cui ho bisogno,questa put parte da storage\app\public,noi gli diciamo di creare la cartella uploads
@@ -127,6 +166,13 @@ class PostController extends Controller
         // $post->save();
         // se $post contiene già i id significa che esisteva già e fa update altrimenti fa inserisci
         $post->update($dati);
+        if (!empty($dati['tag_id'])) {
+            $post->tags()->sync($dati['tag_id']);
+        } else  {
+            $post->tags()->sync([]);
+        }
+
+
         return redirect()->route('admin.posts.index');
     }
 
@@ -139,7 +185,15 @@ class PostController extends Controller
     public function destroy(Post $post)
     {
         $post_image = $post->cover_image;
-        Storage::delete($post_image);
+        if (!empty($post_image)) {
+            Storage::delete($post_image);
+        }
+
+
+        // equivale a mettere cascade delete nel database
+        if ($post->tags->isNotEmpty()) {
+            $post->tags()->sync([]);
+        }
         $post->delete();
         return redirect()->route('admin.posts.index');
     }
